vendor:xosp: make Chromium optional
We noticed that Chromium doesn't fits well on lower end devices leaving the system partition with very less space left. Unable to flash GApps,XOSPApps or other flashable system zips. Therefore leave this as a choice for the maintainer to rather include it or not.

In order to use Chromium in your build you will have to include this flag in your xosp.mk for arm devices:

USE_CHROMIUM := true

For x86 architecture instead there are 2 flags to be added:

USE_CHROMIUM := true

CHROMIUM_X86 := true

Signed-off-by: Calin <nilac8991@gmail.com>