envsetup.sh: Remove unnecessary redirections
'2&>1 >/dev/null' is incorrect when the path doesn't exist at all, it
incorrectly creates a file named '1' using redirect >1 and creates a
folder named '2'.
'2&>1 >/dev/null' only applies when the path already exits and to shut
the noise. Remove the redirects, keep this commit as reference as why it
happened and cleanup the whole xosp_essentials configuration.

Change-Id: Iaa7974cf80b410ff476c642282658acb2ccdc306
Signed-off-by: Kalpaj Chaudhari <daedroza@gmail.com>